Refactor sportbenefit web tests to object-driven suites
