+++ test_new_web_site_sb/test_forms_sb.py
@@ -1,43 +1,15 @@
 # -*- coding: utf-8 -*-
 import time
-from urllib.parse import urlparse
 
 import allure
 import pytest
 
 from pages.new_web_site_sb.companies_sb import CompaniesPageSb
+from pages.new_web_site_sb.contact_forms_sb import ContactFormsSb
 from pages.new_web_site_sb.contacts_sb import ContactsPageSb
 from pages.new_web_site_sb.header_sb import HeaderPageSb
 from pages.new_web_site_sb.main_page_sb import MainPageSb
 from pages.new_web_site_sb.partners_sb import PartnersPageSb
-from test_new_web_site_sb.form_helpers_sb import (
-    accept_cookie_if_present,
-    assert_modal_send_disabled,
-    assert_modal_send_enabled,
-    clear_real_network_probe,
-    clear_submit_spy,
-    close_modal_if_open,
-    ensure_modal_checkbox_checked,
-    fill_inline_contacts_invalid,
-    fill_inline_contacts_values,
-    fill_modal_input,
-    fill_modal_textarea,
-    get_contact_post_urls,
-    get_failure_stub_events,
-    get_modal_send_button,
-    inline_submit_button,
-    inline_submit_enabled,
-    inline_success_state,
-    inline_validation_messages,
-    install_network_failure_stub,
-    install_real_network_probe,
-    install_submit_spy,
-    modal_validation_messages,
-    open_modal_by_cta_text,
-    submit_modal_and_collect_contact_urls,
-    wait_for_contact_network_events,
-    wait_for_modal_submit_feedback,
-)
 
 
 MOBILE_VIEWPORTS = [
@@ -45,10 +17,6 @@ MOBILE_VIEWPORTS = [
     (390, 844),
     (768, 1024),
 ]
-
-SEND_BUTTON_XPATH = (
-    "//div[contains(@class,'modal')]//button[.//span[normalize-space()='Send'] or normalize-space()='Send']"
-)
 
 POST_SUBMIT_FORM_CASES = [
     {
@@ -159,27 +127,7 @@ LIVE_UI_FORM_CASES = [
 ]
 
 
-def _fill_valid_modal_form_post_submit(driver, kind: str, suffix: str):
-    fill_modal_input(driver, "Name", f"QA UX {suffix}")
-    fill_modal_input(driver, "Enter phone number", "+357 99 11 22 55")
-
-    if kind == "ask_question":
-        fill_modal_input(driver, "qwerty@allsports.by", f"qa.ux.ask.{suffix}@example.com")
-        fill_modal_textarea(driver, f"UX question {suffix}")
-    elif kind == "become_partner":
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", f"qa.ux.partner.{suffix}@example.com")
-        fill_modal_input(driver, "Facility name", f"QA UX Facility {suffix}")
-        fill_modal_input(driver, "Enter the city", "Larnaca")
-    else:
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", f"qa.ux.offer.{suffix}@example.com")
-        fill_modal_input(driver, "Company", f"QA UX Company {suffix}")
-        fill_modal_input(driver, "Enter the city", "Limassol")
-
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-
-def _open_target_page_and_modal_for_validation(driver, form_case):
+def _open_validation_page_and_modal(driver, form_case):
     if form_case["page"] == "main":
         page = MainPageSb(driver)
     else:
@@ -187,82 +135,15 @@ def _open_target_page_and_modal_for_validation(driver, form_case):
 
     page.open()
     page.accept_cookie_consent()
-    open_modal_by_cta_text(
-        driver,
-        form_case["cta"],
-        required_placeholder=form_case["required_placeholder"],
-    )
-
-
-def _fill_modal_form_for_validation(driver, form_kind: str, invalid_case: str):
-    email = "qa.matrix@example.com"
-    phone = "+357 99 11 22 33"
-
-    if invalid_case == "invalid_email":
-        email = "bad"
-    if invalid_case == "invalid_phone":
-        phone = "123"
-
-    fill_modal_input(driver, "Name", "QA Matrix")
-    fill_modal_input(driver, "Enter phone number", phone)
-
-    if form_kind == "ask_question":
-        fill_modal_input(driver, "qwerty@allsports.by", email)
-        fill_modal_textarea(driver, "Validation matrix question")
-    elif form_kind == "become_partner":
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", email)
-        fill_modal_input(driver, "Facility name", "QA Matrix Facility")
-        fill_modal_input(driver, "Enter the city", "Nicosia")
-    else:
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", email)
-        fill_modal_input(driver, "Company", "QA Matrix Company")
-        fill_modal_input(driver, "Enter the city", "Limassol")
-
-    if invalid_case != "without_consent":
-        ensure_modal_checkbox_checked(driver)
-
-
-def _require_staging_base_for_live_forms(request):
-    base_url = request.config.getoption("--base-url").rstrip("/")
-    parsed = urlparse(base_url)
-    host = (parsed.netloc or "").lower()
-
-    if "sportbenefit" not in host or "staging" not in host:
-        pytest.skip(
-            "UI live form E2E is allowed only on staging SportBenefit hosts. "
-            f"Current --base-url host: {host or '<empty>'}"
-        )
-
-    return f"{parsed.scheme}://{parsed.netloc}".rstrip("/")
-
-
-def _fill_modal_form_for_live_case(driver, case_kind: str, suffix: str):
-    assert_modal_send_disabled(driver)
-
-    fill_modal_input(driver, "Name", f"QA Live {suffix}")
-    fill_modal_input(driver, "Enter phone number", "+357 99 11 22 33")
-
-    if case_kind == "ask_question":
-        fill_modal_input(driver, "qwerty@allsports.by", f"qa.live.ask.{suffix}@example.com")
-        fill_modal_textarea(driver, f"Live ask-question check {suffix}")
-    elif case_kind == "become_partner":
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", f"qa.live.partner.{suffix}@example.com")
-        fill_modal_input(driver, "Facility name", f"QA Facility {suffix}")
-        fill_modal_input(driver, "Enter the city", "Nicosia")
-    else:
-        fill_modal_input(driver, "qwerty@sportbenefit.eu", f"qa.live.offer.{suffix}@example.com")
-        fill_modal_input(driver, "Company", f"QA Company {suffix}")
-        fill_modal_input(driver, "Enter the city", "Limassol")
-
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
+    forms = ContactFormsSb(driver)
+    forms.open_modal_by_cta_text(form_case["cta"], required_placeholder=form_case["required_placeholder"])
+    return forms
 
 
 @allure.feature("Header SB")
 @allure.severity("Critical")
 @pytest.mark.release_gate
 def test_header_get_offer_modal_structure_sb(driver):
-    """Проверка структуры модального окна Get an Offer из header."""
     page = HeaderPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -276,7 +157,6 @@ def test_header_get_offer_modal_structure_sb(driver):
 @allure.severity("Critical")
 @pytest.mark.release_gate
 def test_main_page_offer_modal_flow_sb(driver):
-    """Проверка полного flow модального окна Get an Offer на главной странице."""
     page = MainPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -286,7 +166,6 @@ def test_main_page_offer_modal_flow_sb(driver):
 @allure.feature("Contacts SB")
 @allure.severity("Critical")
 def test_contacts_page_basics_sb(driver):
-    """Проверка базового открытия страницы Contacts и структуры формы."""
     page = ContactsPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -297,7 +176,6 @@ def test_contacts_page_basics_sb(driver):
 @allure.feature("Companies SB")
 @allure.severity("Critical")
 def test_companies_offer_modal_sb(driver):
-    """Проверка открытия и структуры модального окна Get an Offer на странице Companies."""
     page = CompaniesPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -308,7 +186,6 @@ def test_companies_offer_modal_sb(driver):
 @allure.feature("Partners SB")
 @allure.severity("Critical")
 def test_partners_become_partner_modal_sb(driver):
-    """Проверка открытия и структуры модального окна Become a Partner."""
     page = PartnersPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -320,64 +197,61 @@ def test_partners_become_partner_modal_sb(driver):
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_main_get_offer_form_submit_endpoint_sb(driver):
-    """Проверка отправки формы Get an Offer с главной на endpoint /contact/get_offer."""
     page = MainPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
-
     page.open_get_offer_modal()
-    assert_modal_send_disabled(driver)
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.bot@example.com")
-    fill_modal_input(driver, "Company", "QA Company")
-    fill_modal_input(driver, "Enter the city", "Limassol")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.bot@example.com")
+    forms.fill_modal_input("Company", "QA Company")
+    forms.fill_modal_input("Enter the city", "Limassol")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/get_offer" in url for url in urls), f"get_offer endpoint was not called. Captured: {urls}"
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Main SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_main_ask_question_form_submit_endpoint_sb(driver):
-    """Проверка отправки формы Ask Us a Question с главной на endpoint /contact/ask_question."""
     page = MainPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Ask Us a Question", required_placeholder="qwerty@allsports.by")
 
-    open_modal_by_cta_text(driver, "Ask Us a Question", required_placeholder="qwerty@allsports.by")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@allsports.by", "qa.ask@example.com")
+    forms.fill_modal_textarea("Question from automated tests")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@allsports.by", "qa.ask@example.com")
-    fill_modal_textarea(driver, "Question from automated tests")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/ask_question" in url for url in urls), (
         f"ask_question endpoint was not called. Captured: {urls}"
     )
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Contacts SB")
 @allure.severity("Normal")
 @pytest.mark.form_submission
 def test_contacts_invalid_validation_sb(driver):
-    """Проверка валидации невалидного телефона и email на странице Contacts."""
     page = ContactsPageSb(driver)
     page.open()
     page.accept_cookie_consent()
@@ -388,30 +262,26 @@ def test_contacts_invalid_validation_sb(driver):
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_contacts_inline_form_blocks_invalid_submit_sb(driver):
-    """Проверка, что inline-форма блокирует отправку при невалидных данных."""
     page = ContactsPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    forms.fill_inline_contacts_invalid()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-
-    fill_inline_contacts_invalid(driver)
-
-    submit = inline_submit_button(driver)
+    submit = forms.inline_submit_button()
     if submit.is_enabled() and submit.get_attribute("disabled") is None:
-        driver.execute_script("arguments[0].click();", submit)
+        forms.click_inline_submit()
 
-    messages = inline_validation_messages(driver)
-    urls = get_contact_post_urls(driver)
+    messages = forms.inline_validation_messages()
+    urls = forms.get_contact_post_urls()
 
     assert not urls, f"Inline invalid form should not call contact endpoints. Captured: {urls}"
-    assert not inline_submit_enabled(driver), "Inline submit should remain disabled for invalid values"
+    assert not forms.inline_submit_enabled(), "Inline submit should remain disabled for invalid values"
     if messages:
-        assert any(
-            "invalid" in m.lower() or "format" in m.lower() or "valid email" in m.lower()
-            for m in messages
-        ), (
+        assert any("invalid" in m.lower() or "format" in m.lower() or "valid email" in m.lower() for m in messages), (
             f"Unexpected validation texts: {messages}"
         )
 
@@ -420,36 +290,33 @@ def test_contacts_inline_form_blocks_invalid_submit_sb(driver):
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_contacts_inline_form_submit_endpoint_and_success_state_sb(driver):
-    """Проверка endpoint и успешного состояния после отправки inline-формы Contacts."""
     page = ContactsPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
-
-    fill_inline_contacts_values(
-        driver,
+    forms.fill_inline_contacts_values(
         name="QA Inline",
         phone="+357 99 11 22 33",
         email="qa.inline@example.com",
         company="Inline Company",
     )
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
 
-    if not inline_submit_enabled(driver):
+    if not forms.inline_submit_enabled():
         pytest.xfail(
             "Inline /contacts form submit stayed disabled after valid input. "
             "Endpoint/success-state cannot be validated until page-side gate is fixed."
         )
 
-    submit = inline_submit_button(driver)
-    driver.execute_script("arguments[0].click();", submit)
-
-    urls = get_contact_post_urls(driver)
+    forms.click_inline_submit()
+    urls = forms.get_contact_post_urls()
     assert any("/contact/get_offer" in url for url in urls), (
         f"Inline contacts form did not call get_offer endpoint. Captured: {urls}"
     )
-    assert inline_success_state(driver), (
+    assert forms.inline_success_state(), (
         "Inline contacts form request was sent, but no observable completion signal "
         "(reset/disabled state) was detected"
     )
@@ -459,202 +326,195 @@ def test_contacts_inline_form_submit_endpoint_and_success_state_sb(driver):
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_contacts_get_offer_modal_submit_endpoint_sb(driver):
-    """Проверка отправки модальной формы Get an Offer на endpoint /contact/get_offer."""
     page = ContactsPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Get an Offer", required_placeholder="Company")
 
-    open_modal_by_cta_text(driver, "Get an Offer", required_placeholder="Company")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.contacts.offer@example.com")
+    forms.fill_modal_input("Company", "QA Contacts")
+    forms.fill_modal_input("Enter the city", "Limassol")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.contacts.offer@example.com")
-    fill_modal_input(driver, "Company", "QA Contacts")
-    fill_modal_input(driver, "Enter the city", "Limassol")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/get_offer" in url for url in urls), f"get_offer endpoint was not called. Captured: {urls}"
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Contacts SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_contacts_ask_question_modal_submit_endpoint_sb(driver):
-    """Проверка отправки модальной формы Ask Us a Question на endpoint /contact/ask_question."""
     page = ContactsPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Ask Us a Question", required_placeholder="qwerty@allsports.by")
 
-    open_modal_by_cta_text(driver, "Ask Us a Question", required_placeholder="qwerty@allsports.by")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@allsports.by", "qa.contacts.ask@example.com")
+    forms.fill_modal_textarea("Question from contacts page")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@allsports.by", "qa.contacts.ask@example.com")
-    fill_modal_textarea(driver, "Question from contacts page")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/ask_question" in url for url in urls), (
         f"ask_question endpoint was not called. Captured: {urls}"
     )
-    assert not modal_validation_messages(driver), "Ask-question modal should not show validation errors on valid submit"
-
-    close_modal_if_open(driver)
+    assert not forms.modal_validation_messages(), "Ask-question modal should not show validation errors on valid submit"
+    forms.close_modal_if_open()
 
 
 @allure.feature("Companies SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_companies_get_offer_submit_endpoint_sb(driver):
-    """Проверка отправки формы Get an Offer на корректный endpoint /contact/get_offer."""
     page = CompaniesPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
-
     page.open_get_offer_modal()
-    assert_modal_send_disabled(driver)
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.companies@example.com")
-    fill_modal_input(driver, "Company", "QA Companies")
-    fill_modal_input(driver, "Enter the city", "Limassol")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.companies@example.com")
+    forms.fill_modal_input("Company", "QA Companies")
+    forms.fill_modal_input("Enter the city", "Limassol")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/get_offer" in url for url in urls), f"get_offer endpoint was not called. Captured: {urls}"
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Companies SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_companies_ask_question_submit_endpoint_sb(driver):
-    """Проверка отправки формы Ask Us a Question на endpoint /contact/ask_question."""
     page = CompaniesPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Ask Us a Question", required_placeholder="qwerty@allsports.by")
 
-    open_modal_by_cta_text(driver, "Ask Us a Question", required_placeholder="qwerty@allsports.by")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@allsports.by", "qa.ask.companies@example.com")
+    forms.fill_modal_textarea("Question from companies page")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@allsports.by", "qa.ask.companies@example.com")
-    fill_modal_textarea(driver, "Question from companies page")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/ask_question" in url for url in urls), (
         f"ask_question endpoint was not called. Captured: {urls}"
     )
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Partners SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_partners_get_offer_submit_endpoint_sb(driver):
-    """Проверка отправки формы Get an Offer на странице Partners."""
     page = PartnersPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Get an Offer", required_placeholder="Company")
 
-    open_modal_by_cta_text(driver, "Get an Offer", required_placeholder="Company")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.partners.offer@example.com")
+    forms.fill_modal_input("Company", "QA Partners")
+    forms.fill_modal_input("Enter the city", "Nicosia")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.partners.offer@example.com")
-    fill_modal_input(driver, "Company", "QA Partners")
-    fill_modal_input(driver, "Enter the city", "Nicosia")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/get_offer" in url for url in urls), f"get_offer endpoint was not called. Captured: {urls}"
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Partners SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_partners_become_partner_submit_endpoint_sb(driver):
-    """Проверка отправки формы Become a Partner на endpoint /contact/become_partner."""
     page = PartnersPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
-
     page.open_become_partner_modal()
-    assert_modal_send_disabled(driver)
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.partners.partner@example.com")
-    fill_modal_input(driver, "Facility name", "QA Facility")
-    fill_modal_input(driver, "Enter the city", "Larnaca")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.partners.partner@example.com")
+    forms.fill_modal_input("Facility name", "QA Facility")
+    forms.fill_modal_input("Enter the city", "Larnaca")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/become_partner" in url for url in urls), (
         f"become_partner endpoint was not called. Captured: {urls}"
     )
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("Partners SB Forms")
 @allure.severity("Critical")
 @pytest.mark.form_submission
 def test_partners_ask_question_submit_endpoint_sb(driver):
-    """Проверка отправки формы Ask Us a Question на странице Partners."""
     page = PartnersPageSb(driver)
+    forms = ContactFormsSb(driver)
+
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Ask Us a Question", required_placeholder="qwerty@allsports.by")
 
-    open_modal_by_cta_text(driver, "Ask Us a Question", required_placeholder="qwerty@allsports.by")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Bot")
+    forms.fill_modal_input("Enter phone number", "+357 00 00 00 00")
+    forms.fill_modal_input("qwerty@allsports.by", "qa.partners.ask@example.com")
+    forms.fill_modal_textarea("Question from partners page")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Bot")
-    fill_modal_input(driver, "Enter phone number", "+357 00 00 00 00")
-    fill_modal_input(driver, "qwerty@allsports.by", "qa.partners.ask@example.com")
-    fill_modal_textarea(driver, "Question from partners page")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    urls = submit_modal_and_collect_contact_urls(driver)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    urls = forms.submit_modal_and_collect_contact_urls()
     assert any("/contact/ask_question" in url for url in urls), (
         f"ask_question endpoint was not called. Captured: {urls}"
     )
-
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("SB Validation Matrix")
@@ -663,30 +523,21 @@ def test_partners_ask_question_submit_endpoint_sb(driver):
 @pytest.mark.parametrize("form_case", VALIDATION_FORM_CASES, ids=[c["form_kind"] for c in VALIDATION_FORM_CASES])
 @pytest.mark.parametrize("invalid_case", INVALID_CASES)
 def test_modal_validation_matrix_blocks_invalid_submit_sb(driver, form_case, invalid_case):
-    """Проверка матрицы валидации: невалидные данные не должны проходить как валидные."""
-    _open_target_page_and_modal_for_validation(driver, form_case)
+    forms = _open_validation_page_and_modal(driver, form_case)
+    forms.fill_modal_form_for_validation(form_case["form_kind"], invalid_case)
 
-    _fill_modal_form_for_validation(driver, form_case["form_kind"], invalid_case)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-
-    send_btn = get_modal_send_button(driver)
-    send_disabled_before_click = (not send_btn.is_enabled()) or (send_btn.get_attribute("disabled") is not None)
+    send_disabled_before_click = forms.modal_send_disabled()
     if not send_disabled_before_click:
-        driver.execute_script("arguments[0].click();", send_btn)
+        forms.click_modal_send()
         time.sleep(0.8)
 
-    urls = get_contact_post_urls(driver)
-    send_buttons_after = driver.find_elements("xpath", SEND_BUTTON_XPATH)
-    modal_closed_after_submit = not send_buttons_after
-    send_disabled_after_click = True
-    if send_buttons_after:
-        send_btn_after = send_buttons_after[0]
-        send_disabled_after_click = (not send_btn_after.is_enabled()) or (
-            send_btn_after.get_attribute("disabled") is not None
-        )
-    errors = modal_validation_messages(driver)
+    urls = forms.get_contact_post_urls()
+    modal_closed_after_submit = not forms.modal_is_open()
+    send_disabled_after_click = forms.modal_send_disabled()
+    errors = forms.modal_validation_messages()
 
     if invalid_case == "without_consent":
         assert not urls, (
@@ -708,7 +559,7 @@ def test_modal_validation_matrix_blocks_invalid_submit_sb(driver, form_case, inv
                 f"client-side blocking signal"
             )
 
-    close_modal_if_open(driver)
+    forms.close_modal_if_open()
 
 
 @allure.feature("SB Forms UX")
@@ -716,18 +567,17 @@ def test_modal_validation_matrix_blocks_invalid_submit_sb(driver, form_case, inv
 @pytest.mark.form_submission
 @pytest.mark.parametrize("case", POST_SUBMIT_FORM_CASES, ids=[c["id"] for c in POST_SUBMIT_FORM_CASES])
 def test_modal_submit_success_feedback_sb(driver, case):
-    """Проверка post-submit поведения формы при успешной отправке."""
+    forms = ContactFormsSb(driver)
     suffix = str(int(time.time()))
 
-    driver.get(case["url"])
-    open_modal_by_cta_text(driver, case["cta"], required_placeholder=case["required_placeholder"])
-    _fill_valid_modal_form_post_submit(driver, case["kind"], suffix)
+    forms.open_url(case["url"])
+    forms.open_modal_by_cta_text(case["cta"], required_placeholder=case["required_placeholder"])
+    forms.fill_valid_modal_form(case["kind"], suffix)
+    forms.install_submit_spy()
+    forms.clear_submit_spy()
+    forms.click_modal_send()
 
-    install_submit_spy(driver)
-    clear_submit_spy(driver)
-    driver.execute_script("arguments[0].click();", get_modal_send_button(driver))
-
-    feedback = wait_for_modal_submit_feedback(driver, timeout=8.0)
+    feedback = forms.wait_for_modal_submit_feedback(timeout=8.0)
     assert feedback["closed"] or feedback["disabled"] or feedback["messages"], (
         f"No observable success feedback for modal form {case['id']}: {feedback}"
     )
@@ -739,19 +589,19 @@ def test_modal_submit_success_feedback_sb(driver, case):
 @pytest.mark.parametrize("status_code", [400, 500])
 @pytest.mark.parametrize("case", POST_SUBMIT_FORM_CASES, ids=[c["id"] for c in POST_SUBMIT_FORM_CASES])
 def test_modal_submit_backend_failure_feedback_sb(driver, case, status_code):
-    """Проверка поведения формы при ответах backend 4xx/5xx."""
+    forms = ContactFormsSb(driver)
     suffix = f"{status_code}_{int(time.time())}"
 
-    driver.get(case["url"])
-    open_modal_by_cta_text(driver, case["cta"], required_placeholder=case["required_placeholder"])
-    _fill_valid_modal_form_post_submit(driver, case["kind"], suffix)
+    forms.open_url(case["url"])
+    forms.open_modal_by_cta_text(case["cta"], required_placeholder=case["required_placeholder"])
+    forms.fill_valid_modal_form(case["kind"], suffix)
 
-    install_network_failure_stub(driver, status_code=status_code)
-    driver.execute_script("window.__qaFailNet = [];")
-    driver.execute_script("arguments[0].click();", get_modal_send_button(driver))
+    forms.install_network_failure_stub(status_code=status_code)
+    forms.clear_failure_stub_events()
+    forms.click_modal_send()
 
-    feedback = wait_for_modal_submit_feedback(driver, timeout=8.0)
-    events = get_failure_stub_events(driver)
+    feedback = forms.wait_for_modal_submit_feedback(timeout=8.0)
+    events = forms.get_failure_stub_events()
     matched = [
         event for event in events
         if str(event.get("method", "")).upper() == "POST" and case["endpoint"] in str(event.get("url", ""))
@@ -771,23 +621,19 @@ def test_modal_submit_backend_failure_feedback_sb(driver, case, status_code):
 @pytest.mark.form_submission
 @pytest.mark.parametrize("case", LIVE_UI_FORM_CASES, ids=[c["id"] for c in LIVE_UI_FORM_CASES])
 def test_modal_forms_live_e2e_submit_and_response_sb(request, driver, case):
-    """Проверка live E2E отправки модальных форм и корректного ответа endpoint."""
-    base = _require_staging_base_for_live_forms(request)
+    forms = ContactFormsSb(driver)
+    base = forms.require_staging_base_for_live_forms(request)
     stamp = str(int(time.time()))
 
-    driver.get(f"{base}{case['path']}")
-    accept_cookie_if_present(driver)
+    forms.open_url(f"{base}{case['path']}")
+    forms.accept_cookie_if_present()
+    forms.open_modal_by_cta_text(case["cta"], required_placeholder=case["required_placeholder"])
+    forms.fill_modal_form_for_live_case(case["kind"], stamp)
+    forms.install_real_network_probe()
+    forms.clear_real_network_probe()
+    forms.click_modal_send()
 
-    open_modal_by_cta_text(driver, case["cta"], required_placeholder=case["required_placeholder"])
-    _fill_modal_form_for_live_case(driver, case["kind"], stamp)
-
-    install_real_network_probe(driver)
-    clear_real_network_probe(driver)
-
-    send_button = get_modal_send_button(driver)
-    driver.execute_script("arguments[0].click();", send_button)
-
-    events = wait_for_contact_network_events(driver, case["endpoint"], timeout=30.0)
+    events = forms.wait_for_contact_network_events(case["endpoint"], timeout=30.0)
     assert events, f"No real POST event captured for endpoint {case['endpoint']}"
 
     last_event = events[-1]
@@ -801,7 +647,7 @@ def test_modal_forms_live_e2e_submit_and_response_sb(request, driver, case):
         f"Expected API response, got HTML for {case['endpoint']}: {(response_text or '')[:240]}"
     )
 
-    feedback = wait_for_modal_submit_feedback(driver, timeout=12.0)
+    feedback = forms.wait_for_modal_submit_feedback(timeout=12.0)
     assert feedback["closed"] or feedback["disabled"] or feedback["messages"], (
         f"No observable post-submit feedback for case {case['id']}: {feedback}"
     )
@@ -812,28 +658,23 @@ def test_modal_forms_live_e2e_submit_and_response_sb(request, driver, case):
 @pytest.mark.live_api
 @pytest.mark.form_submission
 def test_contacts_inline_form_live_e2e_submit_and_response_sb(request, driver):
-    """Проверка live E2E отправки inline-формы Contacts и ответа endpoint."""
-    base = _require_staging_base_for_live_forms(request)
+    forms = ContactFormsSb(driver)
+    base = forms.require_staging_base_for_live_forms(request)
     stamp = str(int(time.time()))
 
-    driver.get(f"{base}/en-cy/contacts")
-    accept_cookie_if_present(driver)
-
-    fill_inline_contacts_values(
-        driver,
+    forms.open_url(f"{base}/en-cy/contacts")
+    forms.accept_cookie_if_present()
+    forms.fill_inline_contacts_values(
         name=f"QA Inline {stamp}",
         phone="+357 99 11 22 44",
         email=f"qa.inline.{stamp}@example.com",
         company=f"QA Inline Co {stamp}",
     )
+    forms.install_real_network_probe()
+    forms.clear_real_network_probe()
+    forms.click_inline_submit()
 
-    install_real_network_probe(driver)
-    clear_real_network_probe(driver)
-
-    submit = inline_submit_button(driver)
-    driver.execute_script("arguments[0].click();", submit)
-
-    events = wait_for_contact_network_events(driver, "/contact/get_offer", timeout=30.0)
+    events = forms.wait_for_contact_network_events("/contact/get_offer", timeout=30.0)
     assert events, "No real POST event captured for inline /contact/get_offer"
 
     last_event = events[-1]
@@ -841,8 +682,7 @@ def test_contacts_inline_form_live_e2e_submit_and_response_sb(request, driver):
     assert status in (200, 201, 202, 204), (
         f"Unexpected inline response status: {status}, event={last_event}"
     )
-
-    assert inline_success_state(driver), "No observable inline success state after live submit"
+    assert forms.inline_success_state(), "No observable inline success state after live submit"
 
 
 @allure.feature("SB Mobile Regression")
@@ -850,26 +690,25 @@ def test_contacts_inline_form_live_e2e_submit_and_response_sb(request, driver):
 @pytest.mark.pre_release
 @pytest.mark.parametrize("viewport", MOBILE_VIEWPORTS, ids=lambda v: f"{v[0]}x{v[1]}")
 def test_mobile_main_modal_flow_sb(driver, viewport):
-    """Проверка мобильного flow модальной формы на главной странице для разных viewport."""
-    width, height = viewport
-    driver.set_window_size(width, height)
-
     page = MainPageSb(driver)
+    forms = ContactFormsSb(driver)
+    width, height = viewport
+
+    page.set_viewport(width, height)
     page.open()
     page.accept_cookie_consent()
+    forms.open_modal_by_cta_text("Get an Offer", required_placeholder="Company")
 
-    open_modal_by_cta_text(driver, "Get an Offer", required_placeholder="Company")
-    assert_modal_send_disabled(driver)
+    forms.assert_modal_send_disabled()
+    forms.fill_modal_input("Name", "QA Mobile")
+    forms.fill_modal_input("Enter phone number", "+357 99 11 22 66")
+    forms.fill_modal_input("qwerty@sportbenefit.eu", "qa.mobile@example.com")
+    forms.fill_modal_input("Company", "QA Mobile Co")
+    forms.fill_modal_input("Enter the city", "Limassol")
+    forms.ensure_modal_checkbox_checked()
+    forms.assert_modal_send_enabled()
 
-    fill_modal_input(driver, "Name", "QA Mobile")
-    fill_modal_input(driver, "Enter phone number", "+357 99 11 22 66")
-    fill_modal_input(driver, "qwerty@sportbenefit.eu", "qa.mobile@example.com")
-    fill_modal_input(driver, "Company", "QA Mobile Co")
-    fill_modal_input(driver, "Enter the city", "Limassol")
-    ensure_modal_checkbox_checked(driver)
-    assert_modal_send_enabled(driver)
-
-    send_button = get_modal_send_button(driver)
+    send_button = forms.get_modal_send_button()
     assert send_button.is_displayed(), f"Send button is not visible on viewport {width}x{height}"
 
 
@@ -878,22 +717,21 @@ def test_mobile_main_modal_flow_sb(driver, viewport):
 @pytest.mark.pre_release
 @pytest.mark.parametrize("viewport", MOBILE_VIEWPORTS, ids=lambda v: f"{v[0]}x{v[1]}")
 def test_mobile_contacts_inline_form_validation_gate_sb(driver, viewport):
-    """Проверка мобильной валидации и доступности отправки inline-формы Contacts."""
-    width, height = viewport
-    driver.set_window_size(width, height)
-
     page = ContactsPageSb(driver)
+    forms = ContactFormsSb(driver)
+    width, height = viewport
+
+    page.set_viewport(width, height)
     page.open()
     page.accept_cookie_consent()
     page.check_page_opened()
 
-    fill_inline_contacts_values(
-        driver,
+    forms.fill_inline_contacts_values(
         name="QA Mobile Inline",
         phone="+357 99 11 22 77",
         email="qa.mobile.inline@example.com",
         company="QA Mobile Inline Co",
     )
-    assert inline_submit_enabled(driver), (
+    assert forms.inline_submit_enabled(), (
         f"Inline submit should be enabled for valid values on viewport {width}x{height}"
     )
